Add invalid domain checker, move domain class to its own python file
