+++ src/findcdn/cdnEngine/detectCDN/cdn_check.py
@@ -19,35 +19,15 @@ from ipwhois.exceptions import ASNRegistryError
 # Internal Libraries
 from .cdn_config import COMMON, CDNs, CDNs_rev
 from .cdn_err import NoIPaddress
+from findcdn.findcdn_err import InvalidDomain
+from .domain import Domain
+import validators
+
 
 # Global variables
 LIFETIME = 10
 
 
-class Domain:
-    """Domain class allows for storage of metadata on domain."""
-
-    def __init__(
-        self,
-        url: str,
-        ip: List[str] = [],
-        cnames: List[str] = [],
-        cdns: List[str] = [],
-        cdns_by_name: List[str] = [],
-        namsrvs: List[str] = [],
-        headers: List[str] = [],
-        whois_data: List[str] = [],
-    ):
-        """Initialize object to store metadata on domain in url."""
-        self.url = url
-        self.ip = ip
-        self.cnames = cnames
-        self.cdns = cdns
-        self.cdns_by_name = cdns_by_name
-        self.namesrvs = namsrvs
-        self.headers = headers
-        self.whois_data = whois_data
-        self.cdn_present = False
 
 
 class cdnCheck:
@@ -259,6 +239,15 @@ class cdnCheck:
         interactive: bool = False,
     ) -> int:
         """Option to run everything in this library then digest."""
+
+
+        # check if domain is valid before scanning. this should not happen
+        # since main checks for invalid domains, but a sanity check is good
+        # as we will fail if we get a bad domain at this point
+        if validators.domain(dom.url) is not True:
+            raise InvalidDomain(dom.url)
+
+
         # Obtain each attributes data
         self.ip(dom)
         self.cname(dom, timeout)
